creacion de users index component

+++ app/Http/Controllers/CommentController.php
@@ -26,6 +26,10 @@ class CommentController extends Controller
 
     public function destroy(Comment $comment)
     {
+        if ($comment->user_id != Auth::id()) {
+            abort(403);
+        }
+
         $featureId = $comment->feature_id;
 
         $comment->delete();
